Update editing related components
