+++ newdle/client/src/components/App.js
@@ -36,7 +36,7 @@ export default function App() {
             <Route exact path="/new/success" component={CreationSuccessPage} />
             <Route exact path="/mine" component={MyNewdles} />
             <Route exact path="/participating" component={NewdlesParticipating} />
-            <Route path="/newdle/:code/summary" component={SummaryPage} />
+            <Route exact path="/newdle/:code/summary" component={SummaryPage} />
             <Route path="/newdle/:code/edit" component={EditPage} />
             <Route exact path="/newdle/:code/:partcode?" component={AnswerPage} />
             <Route render={() => <div>This page does not exist</div>} />
