--- newdle/client/src/locales/es/messages.js
@@ -1,1 +1,0 @@
-/*eslint-disable*/export const messages={"<0>newdle</0> is a collective meeting scheduling application.":"<0>newdle</0> es una aplicaci\xF3n para programar reuniones colectivamente.","<0>newdle</0> will collect the answers!":"<0>newdle</0> guardar\xE1 las respuestas","Accept all options where I'm available":"Accept all options where I'm available","Add new participants":"Add new participants","Add participant":"Add participant","Advanced options":"Advanced options","All participants answered.":"All participants answered.","Answer newdle":"Answer newdle","Answering on behalf of:":"Answering on behalf of:","Are you sure you want to leave this page without saving?":"Are you sure you want to leave this page without saving?","Attention":"Attention","Awaiting your reply":"Awaiting your reply","Back":"Back","Cancel":"Cancel","Change participants":"Change participants","Change time slots":"Change time slots","Choose the participants":"Choose the participants","Choose the time slots":"Choose the time slots","Choose your participants":"Elige a los participantes","Click to add time slots":"Click to add time slots","Confirm":"Confirm","Confirm changes":"Confirm changes","Copied!":"Copied!","Copy time slots from previous day":"Copy time slots from previous day","Copy to clipboard":"Copy to clipboard","Create event":"Create event","Create your first newdle!":"Create your first newdle!","Create your newdle!":"Create your newdle!","Delete newdle":"Delete newdle","Delete this newdle":"Delete this newdle","Do you really wish to delete this newdle? Please be aware that <0>this action cannot be undone.</0>":"Do you really wish to delete this newdle? Please be aware that <0>this action cannot be undone.</0>","Done!":"Done!","Duration":"Duration","E-mail participants":"E-mail participants","Edit":"Edit","Email address":"Email address","Error occurred":"Error occurred","Finalize your newdle":"Finalize your newdle","Get started":"Comenzar","Go to newdle summary!":"Go to newdle summary!","How does it work?":"\xBFC\xF3mo funciona?","It is not possible to answer this newdle anymore.":"It is not possible to answer this newdle anymore.","Keep list of participants private":"Keep list of participants private","Leave a comment (optional)":"Leave a comment (optional)","Leave a comment...":"Leave a comment...","Log in":"Log in","Log out":"Log out","Logging in...":"Logging in...","Login":"Login","Login required":"Login required","Logout":"Logout","Made at <0>CERN</0>, the place where the web was born.":"Hecho en el <0>CERN</0>, el lugar donde naci\xF3 la web.","Missing search criteria":"Missing search criteria","My newdles":"My newdles","Name":"Name","Newdles I'm in":"Newdles I'm in","Next":"Next","Next step":"Next step","No data":"No data","No options chosen":"No options chosen","No participants selected":"No participants selected","No users matching the criteria":"No users matching the criteria","Nobody has voted yet.":"Nobody has voted yet.","Note that some of your participants did not provide an email address and thus could not be notified!":"Note that some of your participants did not provide an email address and thus could not be notified!","Notify me about new answers":"Notify me about new answers","Once the newdle is created, you will be shown a link which you need to send to anyone you wish to invite.":"Once the newdle is created, you will be shown a link which you need to send to anyone you wish to invite.","Options":"Options","Participants":"Participants","Please enter a title for your event...":"Please enter a title for your event...","Please enter your name...":"Please enter your name...","Please log in again to confirm your identity":"Please log in again to confirm your identity","Search":"Search","Select final date":"Select final date","Select your language":"Select your language","Send your answer":"Send your answer","Set the time slots based on their availability":"Configura las franjas horarias seg\xFAn su disponibilidad","Skip":"Skip","Some of your participants do not have e-mail addresses and will not be contacted:":"Some of your participants do not have e-mail addresses and will not be contacted:","Something when wrong when notifying participants:":"Something when wrong when notifying participants:","The following participants have not voted yet:":"The following participants have not voted yet:","The participants have been notified of the final date.":"The participants have been notified of the final date.","There are no participants yet.":"There are no participants yet.","They will be presented as options to the participants":"They will be presented as options to the participants","This newdle has already finished":"This newdle has already finished","This newdle has been deleted by its creator.":"This newdle has been deleted by its creator.","This newdle has finished":"This newdle has finished","Timeslots":"Timeslots","Timezone":"Timezone","Update your answer":"Update your answer","We are almost there!":"We are almost there!","Welcome to newdle!":"Bienvenido a newdle!","Who are you?":"Who are you?","You are not part of any newdles yet.":"You are not part of any newdles yet.","You can also \"Skip\" this step and simply share the link to your newdle":"You can also \"Skip\" this step and simply share the link to your newdle","You can use it to find out the best dates/times for your meetings.":"Puedes usarlo para conocer los mejores dias/fechas para tus reuniones.","You need to log in to access this page":"You need to log in to access this page","Your answer has been saved!":"Your answer has been saved!","Your newdle was created and invitation e-mails have been sent. You can send the following link to everyone you would like to invite:":"Your newdle was created and invitation e-mails have been sent. You can send the following link to everyone you would like to invite:","Your newdle was created. You can now send the following to everyone you would like to invite:":"Your newdle was created. You can now send the following to everyone you would like to invite:","Your participants will receive an e-mail asking them to register to their preference. Once the newdle is created, you will be shown a link you can share with anyone else you wish to invite.":"Your participants will receive an e-mail asking them to register to their preference. Once the newdle is created, you will be shown a link you can share with anyone else you wish to invite.","Your session expired":"Your session expired","by {author}":["by ",["author"]],"finished":"finished","newdle is Open Source Software":"newdle es Software de c\xF3digo abierto","newdle logo":"newdle logo","ongoing":"ongoing","originally created in the <0>{newdleTz}</0> timezone":["originally created in the <0>",["newdleTz"],"</0> timezone"],"switch back to the <0>{defaultUserTz}</0> timezone":["switch back to the <0>",["defaultUserTz"],"</0> timezone"],"{0, plural, one {# participant will be e-mailed:} other {# participants will be e-mailed:}}":[["0","plural",{one:["#"," participant will be e-mailed:"],other:["#"," participants will be e-mailed:"]}]],"{0, plural, one {+# more} other {+# more}}":[["0","plural",{one:["+","#"," more"],other:["+","#"," more"]}]],"{0, plural, one {You replied with # timeslot} other {You replied with # timeslots}}":[["0","plural",{one:["You replied with ","#"," timeslot"],other:["You replied with ","#"," timeslots"]}]],"{0, plural, one {{1} out of # participant answered} other {{2} out of # participants answered}}":[["0","plural",{one:[["1"]," out of ","#"," participant answered"],other:[["2"]," out of ","#"," participants answered"]}]],"{0, plural, zero {No participants registered} one {# participant registered} other {# participants registered}}":[["0","plural",{zero:"No participants registered",one:["#"," participant registered"],other:["#"," participants registered"]}]],"{0} will take place on:":[["0"]," will take place on:"],"{availableCount, plural, one {# available participant} other {# available participants}}":[["availableCount","plural",{one:["#"," available participant"],other:["#"," available participants"]}]],"{hours, plural, one {{hours} hour} other {{hours} hours}}":[["hours","plural",{one:[["hours"]," hour"],other:[["hours"]," hours"]}]],"{hours}h {minutes}min":[["hours"],"h ",["minutes"],"min"],"{numSlots, plural, one {<0>#</0> slot selected } other {<1>#</1> slots selected } =0 {You haven't selected any slots yet}}":[["numSlots","plural",{one:["<0>","#","</0> slot selected "],other:["<1>","#","</1> slots selected "],0:"You haven't selected any slots yet"}]],"{numberOfTimeslots, plural, one {{numberOfAvailable} out of # option chosen} other {{numberOfAvailable} out of # options chosen}}":[["numberOfTimeslots","plural",{one:[["numberOfAvailable"]," out of ","#"," option chosen"],other:[["numberOfAvailable"]," out of ","#"," options chosen"]}]],"{title} will take place on:":[["title"]," will take place on:"]};
